refactor: move test session creation to Chat component
