+++ frontend/src/components/canvas/ChatCanvas.tsx
@@ -28,7 +28,6 @@ import { deleteNode, initializeFlow, setNodes, setSelectedEdgeId, setSelectedNod
 import { useModeStore } from '../../store/modeStore'
 import { setNodePanelExpanded } from '../../store/panelSlice'
 import { RootState } from '../../store/store'
-import { createTestSession } from '../../utils/api'
 import {
     insertNodeBetweenNodes,
     useAdjustGroupNodesZIndex,
@@ -131,24 +130,6 @@ const ChatCanvasContent: React.FC<ChatCanvasProps> = ({ workflowData, workflowID
     const mode = useModeStore((state) => state.mode)
 
     const { getIntersectingNodes, getNodes, updateNode } = useReactFlow()
-
-    // Add state for test session
-    const [testSessionId, setTestSessionId] = useState<string | null>(null)
-
-    // Create test session when workflow ID is available
-    useEffect(() => {
-        const initTestSession = async () => {
-            if (workflowID) {
-                try {
-                    const session = await createTestSession(workflowID)
-                    setTestSessionId(session.id)
-                } catch (error) {
-                    console.error('Error creating test session:', error)
-                }
-            }
-        }
-        initTestSession()
-    }, [workflowID])
 
     const handlePopoverOpen = useCallback(
         ({
@@ -271,7 +252,9 @@ const ChatCanvasContent: React.FC<ChatCanvasProps> = ({ workflowData, workflowID
             }
 
             // Don't handle keyboard events if we're focused in the chat panel
-            const chatPanelElement = document.querySelector('.h-full.bg-background.dark\\:bg-background\\/80.border-l.border-divider')
+            const chatPanelElement = document.querySelector(
+                '.h-full.bg-background.dark\\:bg-background\\/80.border-l.border-divider'
+            )
             if (chatPanelElement?.contains(document.activeElement)) {
                 return
             }
@@ -532,14 +515,7 @@ const ChatCanvasContent: React.FC<ChatCanvasProps> = ({ workflowData, workflowID
             </div>
 
             {/* Memoized Chat Panel - only re-renders when props change */}
-            {showChat && (
-                <ChatPanel
-                    workflowID={workflowID}
-                    width={chatWidth}
-                    onResizeStart={startResizing}
-                    testSessionId={testSessionId}
-                />
-            )}
+            {showChat && <ChatPanel workflowID={workflowID} width={chatWidth} onResizeStart={startResizing} />}
         </div>
     )
 }
@@ -563,12 +539,10 @@ const ChatPanel = React.memo(
         workflowID,
         width,
         onResizeStart,
-        testSessionId,
     }: {
         workflowID?: string
         width: number
         onResizeStart: (e: React.MouseEvent) => void
-        testSessionId: string | null
     }) => {
         return (
             <>
@@ -584,7 +558,7 @@ const ChatPanel = React.memo(
                     style={{ width: `${width}px` }}
                 >
                     <div className="w-full h-full overflow-hidden">
-                        <Chat workflowID={workflowID} sessionId={testSessionId} />
+                        <Chat workflowID={workflowID} />
                     </div>
                 </div>
             </>
